chore: moved all utils to index.ts
