+++ middlewares/error.ts
@@ -1,4 +1,4 @@
-import { Request, Response } from "express";
+import { Request, Response, NextFunction } from "express";
 import response from "../utils/response";
 import logger from "../utils/logger";
 import { SOMETHING_WENT_WRONG } from "../constants/errors";
@@ -7,7 +7,7 @@ const useErrorHandler = (
   err: Error,
   req: Request,
   res: Response,
-  next: any
+  next: NextFunction
 ) => {
   logger.error(err);
   const isProduction = process.env.NODE_ENV === "production";
