+++ utils/mail.config.ts
@@ -2,7 +2,6 @@ import nodemailer from "nodemailer";
 import { config } from "dotenv";
 config();
 
-// Configure nodemailer
 const transporter = nodemailer.createTransport({
   host: process.env.MAIL_HOST,
   port: 2525,
